# Optimise record serialisation