I've benchmarked a few string concatenations. It turns out, that out of `fmt.Sprintf`, `strings.Builder`, `strings.Builder` with pre-grow and naive `+`-concat, the latter is the fastest. 

I had the assumption that string concat is special cased in the runtime for up to five operands, but possibly that is not true (anymore). 

```
name \ time/op    printf      buffer      pregrow     concat
Serialization-8   490ns ¬± 1%  245ns ¬± 2%  169ns ¬± 1%  109ns ¬± 1%

name \ alloc/op   printf      buffer      pregrow     concat
Serialization-8    160B ¬± 0%   128B ¬± 0%    96B ¬± 0%     0B     

name \ allocs/op  printf      buffer      pregrow     concat
Serialization-8    8.00 ¬± 0%   6.00 ¬± 0%   4.00 ¬± 0%   0.00     
```