@jwkohnen 2020-02-25
benchstat of `+`-concat with different Go version (all darwin/amd64):

```
name \ time/op    go1.11.13   go1.12.17   go1.13.8
Serialization-8   112ns ± 2%  111ns ± 4%  109ns ± 1%

name \ alloc/op   go1.11.13   go1.12.17   go1.13.8
Serialization-8   0.00B       0.00B       0.00B

name \ allocs/op  go1.11.13   go1.12.17   go1.13.8
Serialization-8    0.00        0.00        0.00
```