@dgryski 2020-03-04
The line protocol and tcp write probably want []byte instead of string.  Might be worth updating the return value here.  Also, `*string` is almost never what you want.  A `string` is a two-word struct (data pointer, string length).  It ought to be passed by value.