@grzkv 2020-04-29
@dgryski Yes. Thanks for pointing this out. We should probably migrate the whole codebase to using `[]byte` instead of `string`.