@azhiltsov 2020-04-29
> @dgryski Yes. Thanks for pointing this out. We should probably migrate the whole codebase to using `[]byte` instead of `string`.

Wouldn't this require a string validation then, which is currently comes with the 'string' type?