@grzkv 2020-04-29
@azhiltsov We may have to keep the `string` type in some places for conversions. But the thing is, we already treat `string` as `[]byte` in many places. I will look into this.